PRR-62 [refactor] modify Step1-3 Form component
